Filled in functions makeCacheMatrix and cacheSolve
Added code to the function makeCacheMatrix that creates a matrix which
is really a list containing functions to set the value of the matrix,
get the value of the matrix , set the value of the inverse matrix, and
get the value of the inverse matrix.

Added code to the function cacheSolve to compute and return the inverse
of the matrix returned by makeCacheMatrix.  If the inverse has already
been calculated, cacheSolve returns the inverse from the cache.